Add support for dictionary fields
A new tab was added in Data section, allowing to define a dictionary
structure.
For each dictionary you can define its keys and associated type.
It's not possible for now to set a an array of a specific type (array
 of dictionary, array of number, etc)

Basic functionality for dictionaries is working.
Lots of strings to update + tests to write.
Also TODO: add a DataField with DataType=Dictionary

refs #12093